Vim-test is working
Not perfect but this taught me a trick.

I can open the app dir where the code is then call nvim .. to open nvim
one dir above. This will give the tests a dir of app but scope for the
whole project to edit code in docker-compose.yaml and Dockerfile etc.