feat: trying to fix lsp issue
Right now the lsp is reporting code that is unreachable when in an ssh
session. When looking at the same code in nvim on the host the lsp does
not have this issue.

Don't know if this is a ssh char issue or something is missing from my
config or the docker container is not hosting Pyright properly etc.